pyproject.toml
[tool.poetry]
name = "melvin"
version = "0.1.0"
description = "A Personal Assistantant that will help to manage your life and boost your productivity"
authors = ["Mihirsinh Chauhan"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.46.2"
torch = "^2.5.1"
openai-whisper = "^20240930"
pytube = "^15.0.0"
typer = "^0.13.0"
llama-index-llms-llama-cpp = "^0.2.3"
llama-index-core = "^0.11.23"
llama-index-readers-youtube-metadata = "^0.2.0"
llama-index-readers-youtube-transcript = "^0.2.0"
python-dotenv = "^1.0.1"
llama-index-embeddings-huggingface = "^0.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
